EVG-2009 Fix test-proto failures

+++ subprocess/subtree_linux.go
@@ -36,7 +36,8 @@ func getEnv(pid int) ([]string, error) {
 	return results, nil
 }
 
-func cleanup(_ string, logger grip.Journaler) error {
+func cleanup(key string, logger grip.Journaler) error {
+	myPid := os.Getpid()
 	pids, err := listProc()
 	if err != nil {
 		return err
@@ -47,7 +48,7 @@ func cleanup(_ string, logger grip.Journaler) error {
 		if err != nil {
 			continue
 		}
-		if envHasMarkers(env) {
+		if pid != myPid && envHasMarkers(key, env) {
 			p := os.Process{}
 			p.Pid = pid
 			if err := p.Kill(); err != nil {
